Remove standalone PPR matrix test file as tests are moved to existing test suite
