+++ releasenotes/notes/ppr-rust-matrix-support-15869.rst
@@ -1,0 +1,6 @@
+---
+features:
+  - |
+    Added Rust-side matrix support for ``PauliProductRotationGate`` by reusing
+    the existing Pauli matrix construction logic from the Rust
+    ``SparsePauliOp`` implementation.
